requirements.txt
scikit-learn
django
django_rest_framework